feat: consolidate sections-add-one and sections-update-one into sections-manage
- Merges sections-add-one and sections-update-one into a single sections-manage tool
- Uses natural parameter detection: if id is provided → update, else → add
- For updates: requires id and name
- For creates: requires name and projectId
- Returns proper MCP error format when required parameters are missing
- Reduces tool count from 15 to 14 tools
